python-rag/requirements.txt
fastapi
uvicorn[standard]
llm
llm-embed-jina
qdrant-client
python-dotenv
requests
news-please
selenium
webdrivermanager